[NO-JIRA] Refactor api structure
