+++ helpers/apiStructureWithDescriptions_Extensions.js
@@ -1,0 +1,6 @@
+const apiStructureWithDescriptionsExtensions = apiStructure => {
+  return apiStructure;
+};
+
+export default { apiStructureWithDescriptionsExtensions };
+export { apiStructureWithDescriptionsExtensions };
