+++ helpers/apiStructure_Extensions.js
@@ -1,0 +1,4 @@
+const apiStructureExtensions = {};
+
+export default { apiStructureExtensions };
+export { apiStructureExtensions };
