+++ server/api/actions/apiStructureWithActions_Extensions.js
@@ -1,0 +1,6 @@
+const apiStructureWithActionsExtensions = apiStructure => {
+  return apiStructure;
+};
+
+export default { apiStructureWithActionsExtensions };
+export { apiStructureWithActionsExtensions };
